+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog for NeoFS Contract
 
 ### Added
 - `cmd/dump` app that pulls state and data of contracts from remote networks (#324)
+- `tests/migration` framework for storage migration testing (#324)
 
 ### Changed
 ### Updated